Salary: float to decimal in C#, and decimal to money in SQL Server

+++ SchoolManagement/Entities/Teacher.cs
@@ -9,11 +9,11 @@ namespace SchoolManagementEntity
         public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
-        public float Salary { get; set; }
+        public decimal Salary { get; set; }
         public int Experience { get; set; }
         public bool Director { get; set; }
 
-        public Teacher(int id, string name, string surname, float salary, int experience, bool director)
+        public Teacher(int id, string name, string surname, decimal salary, int experience, bool director)
         {
             Id = id;
             Name = name;
